pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "task-star"
version = "1.0.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A robust tool for automating questionnaire filling with strategy support."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.1",
    "pyyaml>=6.0",
    "customtkinter>=5.2.1",
    "pillow>=10.0.0",
]

[project.scripts]
taskstar = "task_star.main:main"
